Created testclient package and removed integrationtests.utils
