+++ storage/ndb/rest-server/rest-api-server/internal/integrationtests/testclient/encoding.go
@@ -1,0 +1,27 @@
+package testclient
+
+import "encoding/base64"
+
+// Encode KV Filter
+func EncodePkValue(data string, binary bool, colWidth int, padding bool) string {
+	if binary {
+		newData := []byte(data)
+		if padding {
+			length := colWidth
+			if length < len(data) {
+				length = len(data)
+			}
+
+			newData = make([]byte, length)
+			for i := 0; i < length; i++ {
+				newData[i] = 0x00
+			}
+			for i := 0; i < len(data); i++ {
+				newData[i] = data[i]
+			}
+		}
+		return base64.StdEncoding.EncodeToString(newData)
+	} else {
+		return data
+	}
+}
